refactor(mobile): extract Source Control screen into embeddable panel
Split the 2243-line inline screen into MobileSourceControlPanel + focused
Content/Header/Modals/FileRows/BranchDiffPreviewDrawer components and
state/loaders/runners/openers/action-sheet hooks + styles (all under the
line limit). embedded mode drops full-screen chrome and routes back/close
+ the file-open origin:session dismissal through onRequestClose. Route file
is a thin wrapper; full-screen behavior unchanged (U2).
